Change xc_copy to use dedicated regression data folder
In the test xc_copy a file used for a successive output and input
of COPY was directed to folder /tmp/, which may lead to problems
because of undeleted data depending on the system used.
So change this file path to use the path used by data files of
regressions.