Fix make install for vpath installation
Documentation tool used for Postgres-XC documentation compilation
called makesgml is moved to a specific tool repository out of the
documentation folder itself to strengthen independency of this module.
Makefile structure is also reviewed in a couple of places directly
dependant on makesgml like bin/psql.
Some other issues related to the way GTM binaries were installed are
fixed and treated now as consistently as Postgres.