Simplify remote node communication at PREPARE phase
This commit fixes issue 3528952, which made a Coordinator sending
COMMIT messages unnecessarily to remote nodes for read connections.
This caused Datanodes to log a lot of warning messages of the type
"WARNING: there is no transaction in progress". This issue was
reproducible when cluster was under a huge load.

Things are simplified in pgxc_node_remote_prepare and this function
does not manage anymore with COMMIT messages, what it was not meant
to do by the way.

The modifications of this commit revert partially things implemented
in commit 077b37369a4b26b5aa0db7b23667a1803e2af4de but does not
change the process of SELECT FOR UPDATE transactions as connections
using such queries are registed as write connections to remote nodes.
All the test cases introduced by commits 077b373 and 9e66733 are kept.
They pass correctly with this patch, and measurements with DBT-1 are
passing correctly.