Fix a problem with transactions staying open on remote nodes
The origin of the problem was that BEGIN messages were being sent to remote nodes
for all the queries launched in do_query of execRemote.c. However, read-only
queries were registered as read connections to remote nodes, meaning that a
COMMIT message was not sent in this case. This had as direct consequence to keep
transactions open on remote nodes.

For example, in such a test case:
create table tab(a int, b int);
insert into tab values(1,2),(3,4);
begin;
select * from tab;
prepare transaction 'pt2';
select * from tab;

The transaction is begun on local Coordinator and remote nodes, but it is not
committed on remote nodes. The origin of the problem is that it doesn't make
sense to send a BEGIN message to remote nodes when SELECT is launched inside
this transaction block, but it needs only to be sent when the first write operation
is done.
Then, when the second SELECT was launched, the remote Datanodes were complaining
that a transaction was already open.

This fix has been checked with regressions and DBT-1.
It improves the performance output of DBT-1 by 15% for 5 nodes thanks
to all the BEGIN messages not sent to remote nodes.