Block node registration for slave nodes and perform register at a stable state
Performing node registration on GTM  just after postmaster start is dangerous
because the master/slave status of the node is unknown. So node
registration is done once node has reached a stable state, meaning that
Xlog process has confirmed to postmaster that recovery has been completed.

Registration is also blocked for slave nodes. A standby node is created
from a hot backup of its master so master and its slaves will share the
same node name. Two nodes are not authorized to register to GTM with the
same node name to avoid identification confusion in the cluster, so it
looks normal to let the master and the master only identify itself.
A a direct consequence, unregistration too is done only for master node.

Synchronous and asynchronous node replication works correctly with this fix,
as well as node promotion (Ex: "pg_ctl promote -D $DATA").