Fix for bug 3535915: INSERT SELECT for partially replicated tables
Remote DML planning was not taking into account the node list
to be used on a relation when modifying data on it.

This has been enlighten by a test case with INSERT SELECT on
tables replicated defined on a sub-cluster. For example, in the
case of a cluster with 2 Datanodes dn1 and dn2:
CREATE TABLE aa1 (a int) DISTRIBUTE BY REPLICATION TO NODE dn1;
CREATE TABLE aa2 (a int) DISTRIBUTE BY HASH(a);
INSERT INTO aa2 VALUES (1),(2),(3),(4);
INSERT INTO aa1 SELECT * FROM aa2;
The last INSERT SELECT was inserting data on all the nodes instead
of doing it only on Datanode dn1.