Fix for crash at gtm_proxy startup in 32b-linux machines
This fix adds an additional boolean flag in GTM-Proxy thread info
to make it similar to the one used in GTM. The problem which was
happening was that GTM-Proxy tried to use the memory structure of
the thread info of GTM.
Crash problems are solved but we should find in the future a smarter
way to manage GTM and GTM-Proxy memory allocation APIs as they are
now strongly dependant on each other and we may find similar issues
if some new developments are done on GTM and/or GTM-Proxy.

Report from Plexo Rama and patch from Koichi Suzuki.