Refactor regression tests xc_misc, xc_FQS and xc_FQS_join
Those tests are using a similar PL/PGSQL function to create tables with
a node list, being transparent on the cluster structure. This function
was created 3 times. Now this function is created in another test
called xc_create_function launched before all the others.
A table name conflict has been resolved in xc_misc to allow this test
to run in parallel with the other XC-related tests.