Fix for bug 3540784: Incorrect COPY for partially distributed tables
When a table was distributed on a portion of nodes. COPY process
was not correctly choosing the output node during a COPY TO. This
problem has been put into light by a case where table was replicated
without using the first datanode of the cluster. It happened that
there was the same problem for hash/modulo tables as well.
A test case in xc_copy has been added to cover this problem, and
a small portion managing COPY TO in execRemote.c has been cleaned.