Command Id communication protocol between Postgres-XC	nodes
This commit adds a feature to allow Postgres-XC	nodes to communicate
a Command Id to	remote nodes. Remote nodes can also send back a	Command
Id to the Coordinator commanding the transaction.
This has as consequences to solve numerous issues with data view on cursors
as well	as solving an old issue	XC had with INSERT SELECT when INSERT is
done on a child after scanning the parent.
This also allows to correctly increment the command ID on Coordinator in
the case of triggers and/or constraints being fired on remote nodes.

In order to allow nodes	to communicate command ID, and a new message type
'M' is added.
Note: this patch is also fixing some whitespaces in xact.c.

Patch from Abbas Butt, I just added some simplifications, comments,
and finalized the packing.