Set CommandId communication flag on Coordinator only for parent/child INSERT SELECT
In the particular case of an INSERT SELECT where INSERT is done on a child table
based on the data scanned of the parent, it is necessary to communicate the command Id
between Postgres-XC nodes. However in this case it was done for all the nodes and it
is only necessary to set it on local Coordinator.