Bypass ERROR message when IF EXISTS is activated on objects
Default behavior is now to set missing_ok at true to avoid
an ugly "relation" for each non-existing object in catalog.
In case an object does not exist on local node, the DDL query
is not sent to remote nodes
This way of doing fixes 2 issues:
1) Bypass ERROR if IF EXISTS is used
2) Let the local node throw correct NOTICE and ERROR back to
clients,

This commit contains a fix for tests foreign_data and sequence.