Block the use of views in LIKE clauses
Block the creation of tables using views in their LIKE clause.
Views are not created on Datanodes, so this will result in an error.
In order to fix this problem, it looks necessary to transform the query
string of CREATE TABLE into something not using the view definition.
Now Postgres-XC only uses the raw string... There is some work done with
event triggers in 9.3, so it might be possible to use that code to
generate dynamically at utility executir step the DDL query to be sent
to remote nodes. When this is done, this error will be removed.