Correct management of RULE drop and creation on cluster
Prior 9.2, RULE and TRIGGER DROP was included under DropPropertyStmt.
But this has been moved in DropStmt with all the other types of objects.
So it has been necessary to adapt the analysis of objects when rules are
dropped for permanent and temporary objects.

This commit fixes test cases rules and drop_if_exists.