Fix for bug 3556031: pgadmin3 connection error
pgadmin3 parses the output of "SELECT version();" to determine
the server version it is connecting to. For Postgres-XC, this
output result was changed and completed with Postgres-XC and
PostgreSQL version. pgadmin thought that the version of server
was the one of XC, meaning 1.0 for 1.0 stable and 1.1 for master
branch. In consequence to that, SQL queries with very old catalog
were run to server, and of course errors were reported to client,
and pgadmin cannot work correctly.
This is solved with changing back version() to its former output.
An additional function pgxc_function() is added to provide a way
for application to inform about Postgres-XC version and the PostgreSQL
version it is based on.

Change is backpatched to all supported versions.