Refactoring of pg_regress.c due to bugs with O2 compilation
When code was being compiled with O2, initialization and start-up
of GTM was done twice, creating errors in log files.
In consequence, functions related to start and initialization of
PGXC nodes and GTM are unified to clarify code and additional checks
are added to verify that nodes have been correctly started.

Conflicts:
	src/test/regress/pg_regress.c