Fix snapshot management on slave nodes and Datanodes accessed directly by applications
The following things are changed:
- Make a slave node session acquire its snapshot directly from WAL
like PostgreSQL does. We do not try to get a global snapshot from GTM
- Allow a Datanode to get a snapshot directly from GTM when accessed directly
by an external application
- Block run of queries interacting with remote nodes on a slave nodes.
In this case it is necessary to obtain a transaction ID but a node in
recovery cannot do that.

in consequence, the following issues are fixed:
- Allow read queries to run on Coordinator slaves by allowing getting a
snaphot. A slave Coordinator was complaining about the impossibility to get
a transaction ID when fetching a snapshot for all types of queries. This allows
to check if the slave Coordinator is still alive and to monitor its statistics
as long as the query does not involve data located on remote Datanodes.
- Suppression of the annoying WARNING "Cannot get a global snapshot from GTM"
when connecting to a Datanode directly with a psql client. Now in this case
a snapshot is taken directly from GTM. Processes like autovacuum did that already.
- Make slave nodes getting there snapshots directly from WAL and not from GTM.

This is also backpatched to 1.0.