Refactor pgbench code for XC and add documentation
There is an additional mode in pgbench dedicated to Postgres-XC
to help testing easily the write-scalability of the system tested.
By using option -k at initialization, it is possible to force the
tables to be hash-distributed by column bid.
When launching a test, once again using -k allows to enforce write
queries to use an additional condition on bid column to reduce the
node selectivity inside transaction into a single node. This helps
testind the write-scalability of the system by dividing write load
based on a unique key, and reduces the effects of locks due to the
update of pgbench_accounts.
At the same time some old bugs are fixed.