Fix race condition in GXID generation for autovacuum wraparound
When an autovacuum wraparound begins, a global snapshot with a
global GXID is obtained from GTM. After getting this snapshot,
the code checks for wraparound conditions. Before this check, the
lock on GXID generation is released and then taken back. Just after
taking back the lock the GXID is updated to the latest value in cache.
However, during this time it is possible that ShmemVariableCache->nextXid
is updated by another local session that obtained a transaction ID
from GTM. GXID generation concurrency is managed globally at GTM, so we
cannot rely here on the local ShmemVariableCache->nextXid to update the
GXID of autovacuum. Updating it to a wrong value makes is async with GTM
information and such autovacuum is not able to gather a global snapshot
correctly, making the autovacuum session inconsistent with the other global
sessions. This could fatally crash a node under certain conditions like
GXID being updated concurrently by multiple sessions at the same time
during the wraparound condition checks.

This bug was easily reproducible by setting in a fresh cluster the first
GXID value on GTM to a value close to autovacuum_freeze_max_age, and
then run the cluster for a couple of minutes with pgbench for example.