Remove reserved keywords REPLICATION, MODULO, HASH, ROUND and ROBIN
Those keywords were used for clause DISTRIBUTE BY to define a
distribution type. In order to being more flexiblity and extensibility
on this code, a method based on IDENT is instead used.
This fixes directly an issue with ALTER ROLE which was using the
word REPLICATION to allow a user to use the replication protocol.