Reorganize XC planner APIs for shippability evaluation
In order to improve Postgres-XC planner API visibility and transparency, all the
functions that are related to the evaluation of expression shippability to remote
nodes are now gathered into a unique file called pgxcship.c located in PostgreSQL
optimizer.
Most of those functionnalities were originally located in file src/pgxc/planner/planner.c.
This clean-up allows to keep separated the XC planning functions depending on their role
and facilitates code maintenance.
In order to evaluate if an expression or a query is shippable to remote nodes, a simple call
to the new APIs of pgxcship.h is enough.

The file called postgresql_fdw.c is removed, made unnecessary by the exposition of
shippability evaluation functions in new interface.