Remove option strict_statement_checking
This option was used in XC planner to enforce an error in case
other queries than SELECT, INSERT, UPDATE and DELETE went through
XC planner. This was particularly useful for debugging but as now
XC is far more integrated with Postgres code, such options are no
more necessary. We need to rely on vanilla for such things on a
local node.