Support for trigger shippability evaluation and statement triggers
This commit adds support for triggers using FOR EACH STATEMENT. An XC
specific infrastructure is added to evaluate if a trigger on a relation
is shippable to a remote node or not.

This needs to be considered as a first step targetting a complete support
of triggers and constraints in Postgres-XC. The second step of the
implementation will be to improve the synergy of constraints and triggers
for XC-specific features regarding their shippability to remote nodes.

In vanilla, constraints and triggers use the same structure and are
fired during DML execution step if necessary.
However, constraints are implemented in such a way in XC that only the
constraints that can be locally enforced on Datanodes can be created.
For example, a simple restriction is the case of a primary on a hash
distributed table, the primary key must contain the distribution column
to allow its check on remote nodes.
Well, this lacks of extensibility because we need to be able to check
constraint validity and fire triggers also on local Coordinators to
allow all possible types of operations including foreign keys or
fancier things in cluster. For example, in the case of triggers it is
necessary to fire triggers using a non-shippable volatile function on
local Coordinator to insure database integrity. Once those extensions
are in place, a real support for FOR EACH ROW trigger and more global
constraints will be possible by allowing constraints and triggers to
be fired also on Coordinators safely. This will need a proper row-based
fetch mechanism though.

By the way, row-based triggers are disabled for the time being.
Documentation is cleaned up and CREATE/ALTER/DROP TRIGGER commands
are re-allowed. Some test cases for different types of triggers are added
with a new regression test xc_triggers.