Get consistent sequence dump value with pg_dump
In Postgres-XC, the current value cached for a sequence may differ at node
level because of the shared-nothing nature of the cluster. Vanilla pg_dump
has been designed to take directly the current value of sequence from the
cache at node level, which is incorrect in XC and might lead to inconsistencies
if for example a dump is taken on a different node than where nextval has
been lastly called.

Report from Nikkhil Sontakke.
Nikhil proposed a fix which was not entirely consistent with Postgres
by using currval, so I created a solution with nextval to insure that
the correct and consistent cluster-wide value of sequence is taken for its
dump.