The commit has following changes. 1. If each group according to GROUP BY clause resides completely on a single datanode i.e. GROUP BY has distribution column in it for distributed underlying relation or the underlying relation is a replicated relation, finalise the aggregates on the datanodes. There is no need for a covering Group or Agg node in the standard planner. 2. All the optimizations related to GROUP BY and aggregation are now available even if the query has ORDER BY or DISTINCT clauses in it. (Earlier they were not.)
Implementation details
---------------------------------
We push aggregates to the datanodes to get transition results for each group and
then apply combination and finalization phase at the coordinator. This is done
by substituting Aggref nodes in RemoteQuery targetlist (representing results of
the datanodes aggregation) as arguments to the Aggref nodes in Agg plans and
letting set_plan_refs take care of Var substitution. But since this changed the
Aggref nodes on Agg plan it was not possible for the nodes above Agg node to
find matching Aggref nodes and thus set the plan references correctly. That's
why we were not able to use these optimizations in the presence of ORDER BY or
DISTINCT clauses which require nodes to be added on top of Agg node (if any).
Hence, moved this logic of substitution to set_plan_refs() and is applied only
for Agg node.