Refactor code controlling constraint creation evaluation
The code refactored in this commit is present in XC since the beginning of
the project and was outdated. The former code was put inside the utility
deparser to control UNIQUE, PRIMARY KEY, EXCLUDE and FOREIGN KEY creation.
This was a fundamental mistake as this did not allow control of FOREIGN KEY
creation with generic functions that could also be used to evaluate the
shippability of constraints when issuing a DML on a given relation or when
modifying a relation distribution with ALTER TABLE (online data redistribution
feature).
Now all this control is put directly inside Postgres-XC shippability evaluation
code and can be used in a wider way than the previous implementation. All the
error messages are now made generic and refer to the impossibility to create
constraints whose evaluation cannot be enforced to remote nodes.
Postgres-XC does not support yet global constraints, but once it does, it will
be necessary to remove the error messages referring to constraint evaluation
and to use the APIs implemented in this commit not to control the creation of
constraints but to check the shippability of a Query for a given relation
constraint.
The old code was also deeply lacking evaluation of index expressions, so some
control is added about that. A regression test called xc_constraints check the
shippability of constraints created on tables depending on their distribution.
Documentation is also updated, but might need more completion.
The code in this commit checks if redistribution with ALTER TABLE can be done
according to the given constraints already on it for UNIQUE, EXCLUDE and PRIMARY
KEY. For example, you should not be able to change a replicated table with a
primary key to roundrobin you cannot ensure the uniqueness of a key on such
tables so now this is protected properly. However, such checks are still missing
for FOREIGN KEY and REFERENCES.