Refactor and clean-up of relation locator code
Locator is the part in charge of managing, building and referencing
distribution information of tables in catalogs by providing sufficient
functions that backends can refer to.
Some of this code was pretty outdated and a portion of the APIs used
were duplicated, making the code more difficult to apprehend. Locator
information also included the following information:
- OID of relation it refers to
- Attribute number of distribution key if distribution is value-based
- Column name of distribution key
As column name can be retrieved from system cache by using simply OID
and attribute number, the column name is removed, simplifying the code
by that much. So now all the APIs of locator.c and related only use the
attribute number of a relation when referencing the distribution key inside
code. This cleanup also tackles some potential bugs related to hash/modulo
management that might have appeared if other value-based distribution would
have been defined in XC. So code is more extensible now.