Index expressions shippability analysis for distribution columns
This commit strengthens determination analysis of unique index expressions
shippability for value-distributed tables.

For example, in the case of a cluster with 2 Datanodes:
CREATE TABLE aa (a int) DISTRIBUTE BY HASH(a);
CREATE UNIQUE INDEX aap ON (abs(a));
INSERT INTO aa (2); -- To Datanode 1
INSERT INTO aa (-2); -- To Datanode 2
The index uniqueness is not ensured here even if the index expression constains
only the distribution column of parent relation. Such expressions are not
shippable, and as XC does not support yet global constraints such index creation
is simply blocked.

Also, if the table has its data located on a single node, its index uniqueness
can be evaluated safely in all the cases, a condition is added to treat that
case.