Fix memory corruption for GTM/backend buffer on backend side
This commit fixes an issue that corrupted data from GTM on Postgres backend for
snapshots and transaction IDs. The origin of the failure was the startup
package of GTM that was not initialized correctly, so the data it contained
could be used as snapshot or GXID data in very particular situations like when
a vacuum freeze is kicked on a system table due to the table oldest XID being
too old in accordance to settings like autovacuum_freeze_max_age and
autovacuum_freeze_min_age.

In order to reproduce easily the problem, I personnally had to use a cluster
of 5 nodes with autovacuum_freeze_max_age at its minimum value, then wait for
a system table to be automatically vacuumed. Once the vacuum was kicked, the
buffer of 16kb used at backend side to store messages from GTM got corrupted,
having as consequence to generate multiple types of crashes related to memory
allocation (realloc and malloc calls).

It is really surprising that this issue was not reproducible in other
circumstances and could only be spotted with a forced vacuum as it was a pretty
general problem related to GTM communication, and valgrind helped to spot pretty
easily the memory area that was corrupted.