Move GTM configuration APIs in src/gtm/common to new folder src/gtm/config
This is part of the effort to eliminate the duplications in GTM code with things
of PG src/backend. src/gtm/common had a dependency with the binary postgres when
it is built so this dependency with GTM configuration is unnecessary. Even if
GTM should rely on the APIs of guc-file.l for configuration file treatment, for
the time being this permits to split each functionality correctly.