Add documentation for XC-specific query plan tuning options
Those options are usually used by applications developers to debug correctly
queries relying on the Postgres server framework like EXPLAIN output. In the
case of PG, all the planner-related parameters are documented and listed in
postgresql.conf. Surprisingly, it was not the case of XC-specific parameters
which are:
- enable_fast_query_shipping, to control FQS planner
- enable_remotejoin, to control remote join planning
- enable_remotegroup, to control remote GROUP BY planning
- enable_remotelimit, to control remote LIMIT planning
Like for PG parameters, all those parameters are enabled by default.
Documentation is also provided, and there corresponding entry is added to
postgresql.conf.