Refactor Makefile of GTM binaries and remove unnecessary build dependencies
All the Makefiles of src/gtm are changed to fit in a better way the Postgres
style used inside src/backend. It is now possible to invocate make directly
inside src/gtm/main or src/gtm/proxy to generate only the related binary. there
are not anymore annoying dependencies with the order folders have to be
compiled. Some strange dependencies are also fixed in this commit. A couple of
functions related only to GTM-Standby using structures defined in gtm/main were
located in src/gtm/recovery, which is also a folder GTM-Proxy has a dependency
to, so this was leading to unncessary dependencies with the GTM main and
GTM-proxy main messing up all the building process. Those GTM-only files are
moved to gtm/main to avoid such problems.
Prior to this commit, src/gtm was creating and installing a series of so
libraries that are indeed not necessary for the installation as they are only
server side features. It is not that difficult to reallow them but for the time
being it is better to keep a GTM installation as light as possible.