This commit fixes several bugs:
1. --status does not work
2. Failed to set xc_maintenance_mode. () message is printed.

First of all, now --status is a submode of --verbose.   It was modified
to handle --status as --verbose.   This is a quick fix and should improve
in the next major release.

Second was caused by SET XC_MAINTENANCE_MODE = ON to template0.  It is
not correct and now template0 was skipped.