Enforce autocommit for ALTER DATABASE SET TABLESPACE
ALTER DATABASE can be run inside a transaction block except when the option
SET TABLESPACE is specified, so in this case this was simply failing as cluster
thought that this DDL was running inside an internal transaction block. The fix
here is simply to enforce autocommit when running the query. This is not the
safest way expecially if this query fails on the way but this maintenance
branch can live with that. Current master branch safely takes care of DDLs that
cannot be run inside a transaction block even if multiple nodes are involved,
so this fix is not needed at all there.