This commit fixes the bug 3583921, where KnownAssignedTransactionIds array overflows and slaves stop with error.   This is caused by incontiguous assignment of XID for each node and there could be missing XIDs.
This commit should apply only to REL1_0_STABLE branch because this
simply disables management of KnownAssignedTransactionIds, which
ends up with wrong snapshot for hot standby.

In 1.0.x, hot standby does not work because we have no means to provide
consistent database view to slaves.

For following major releases, KnownAssignedTransactionIds should be handled
correctly for future support of hot standby.