The patch allows xc_maintenance_mode to use DML in execute direct and direct connection to datanodes.
Xc_maintenance_mode can be set to on only by superusers and it cannot
be set to on in postgresql.conf or postgres/pg_ctl command line
parameters.   It can be set only in a session.    This extension
maintains its safetiness and allows (only) DBA to cleanup XC cluster
state without reconfiguration.

	modified:   doc-xc/src/sgml/config.sgmlin
	modified:   src/backend/access/transam/xact.c
	modified:   src/backend/parser/analyze.c