This is a fix for -w option in gtm_ctl start, but Nikhil Sontakke.
-------->8---------------->8----------
PFA, a patch to fix "gtm_ctl -w" behavior in GIT HEAD. I also tested
"-w -t nsecs" behavior and it seems to work as well. This patch can be
easily be backported to 1.0 if desired. I hope all these obnoxious gtm
startup issues are resolved now. Scripting was a bit painful because
of these issues.
--------8<----------------8<----------

	modified:   src/bin/gtm_ctl/gtm_ctl.c