Add release notes of 1.0.2 in master branch
Such notes need to be kept in sync in all the upstream branches of the release
that is concerned. For example 1.0.X notes need currently to be updated in 1.0
stable and master. If in the future 1.1 stable is created, the release notes
of 1.0 will be added to it as well as on master. On the contrary, as downstream
notes do not need to be updated, 1.1 notes will not be included in 1.0 stable.

This helps in keeping a consistent release note behavior for all the branches.