This is the first official commit of pgxc_ctl utility in C.
Source directory is /contrib because
1) pgxc_ctl does not support every variety of XC cluster which you can
configure with XC core, and
2) pgxc_ctl makes a couple of assumption (user name, ssh connection and
database owner) which is not necessary in XC core.

They are announced early in the April, 2013 in the mailing list.

Document will be committed later.  Configuration file is almost compatible
with bash version.   You can still write the configuration with bash
script.

Remaining work fir 1.1 is to add a feature to add/remove node.   Abbas
is working on the core feature.

Before the feature release