Support for ROW triggers in Postgres-XC. This includes BEFORE/AFTER/INSTEAD row triggers. Statement trigger support was already implemented.
There were two independent changes; one to support BEFORE ROW triggers, and the
other to support AFTER ROW triggers. INSTEAD triggers worked w/o any changes.

BEFORE ROW triggers:

When a trigger is to be executed from coordinator, the OLD and NEW row is not
available. It needs to be fetched from the DML source plan.

We have been using the DML source plan slot directly to generate the data for
BINDing the parameter values for UPDATE. Now for BEFORE ROW triggers, the data
of NEW row is the one that is to be BOUND. And this NEW tuple slot
does not have the ctid and node_id junk attribute values. So we need
to generate the BIND data using the data from NEW row and the junk
attributes from the source data.
Similarly there are changes in the way we generate the parameters
markers for the remote UPDATE statement.

AFTER ROW triggers:

AFTER ROW triggers are queued for each of the rows processed, and are then
executed at then end of statement. So we need to store the OLD and NEW rows in
memory until the end of statement or until the end of transaction in case of
deferred constraint triggers. The idea is to use tuplestore for saving them.

TODOs:

New feature tests to be added to regression suite.
COPY with INSERT triggers is not working.
A mix of shippable and non-shippable triggers for the same table has issues. Fix
them.
More unit testing.